# DOCS/api-swagger.json
{
   "openapi":"3.0.1",
   "info":{
      "title":"OpenAPI definition",
      "version":"v0"
   },
   "servers":[
      {
         "url":"http://localhost:8080",
         "description":"Generated server url"
      }
   ],
   "paths":{
      "/pessoa/criar":{
         "post":{
            "tags":[
               "Person API"
            ],
            "summary":"Create a new person",
            "operationId":"criarPessoa",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/PessoaRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponsePessoa"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/sacar":{
         "post":{
            "tags":[
               "Management Account API"
            ],
            "summary":"Withdraw value from account",
            "operationId":"sacarValor",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/SaqueRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseTransacao"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/depositar":{
         "post":{
            "tags":[
               "Management Account API"
            ],
            "summary":"Deposit in an account",
            "operationId":"depositarValor",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/DepositoRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseTransacao"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/criar":{
         "post":{
            "tags":[
               "Management Account API"
            ],
            "summary":"Create a new account",
            "operationId":"criarConta",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/ContaRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseConta"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/bloquear/{idConta}":{
         "post":{
            "tags":[
               "Management Account API"
            ],
            "summary":"Blocks an account",
            "operationId":"bloquearConta",
            "parameters":[
               {
                  "name":"idConta",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseConta"
                        }
                     }
                  }
               }
            }
         }
      },
      "/health":{
         "get":{
            "tags":[
               "Endpoints"
            ],
            "summary":"Health check endpoint",
            "operationId":"health",
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/saldo/{idConta}":{
         "get":{
            "tags":[
               "Management Account API"
            ],
            "summary":"Get account balance",
            "operationId":"obterSaldo",
            "parameters":[
               {
                  "name":"idConta",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseBigDecimal"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/extrato":{
         "get":{
            "tags":[
               "Management Account API"
            ],
            "summary":"Get account transaction extract",
            "operationId":"extratoConta",
            "parameters":[
               {
                  "name":"idConta",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               },
               {
                  "name":"dataInicial",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"string"
                  }
               },
               {
                  "name":"dataFinal",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseListTransacao"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "PessoaRequest":{
            "required":[
               "cpf",
               "dataNascimento",
               "nome"
            ],
            "type":"object",
            "properties":{
               "nome":{
                  "type":"string",
                  "description":"Person name",
                  "example":"Everton M"
               },
               "cpf":{
                  "pattern":"[0-9]",
                  "type":"string",
                  "description":"Person CPF",
                  "example":"12345678900"
               },
               "dataNascimento":{
                  "pattern":"^(3[01]|[12][0-9]|0[1-9])/(1[0-2]|0[1-9])/[0-9]{4}$",
                  "type":"string",
                  "description":"Birth date in pattern DD/MM/AAAA",
                  "example":"22/09/1987"
               }
            }
         },
         "Pessoa":{
            "type":"object",
            "properties":{
               "idPessoa":{
                  "type":"integer",
                  "format":"int32"
               },
               "nome":{
                  "type":"string"
               },
               "cpf":{
                  "type":"string"
               },
               "dataNascimento":{
                  "type":"string",
                  "format":"date-time"
               }
            }
         },
         "ResponsePessoa":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "$ref":"#/components/schemas/Pessoa"
               }
            }
         },
         "SaqueRequest":{
            "required":[
               "idConta",
               "valor"
            ],
            "type":"object",
            "properties":{
               "idConta":{
                  "type":"integer",
                  "description":"Account ID",
                  "format":"int32",
                  "example":3
               },
               "valor":{
                  "type":"number",
                  "description":"Withdraw amount value",
                  "example":15.0
               }
            }
         },
         "ResponseTransacao":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "$ref":"#/components/schemas/Transacao"
               }
            }
         },
         "Transacao":{
            "type":"object",
            "properties":{
               "idTransacao":{
                  "type":"integer",
                  "format":"int32"
               },
               "idConta":{
                  "type":"integer",
                  "format":"int32"
               },
               "valor":{
                  "type":"number"
               },
               "dataTransacao":{
                  "type":"string",
                  "format":"date-time"
               },
               "tipoTransacao":{
                  "type":"integer",
                  "format":"int32"
               }
            }
         },
         "DepositoRequest":{
            "required":[
               "idConta",
               "valor"
            ],
            "type":"object",
            "properties":{
               "idConta":{
                  "type":"integer",
                  "description":"Account ID",
                  "format":"int32",
                  "example":3
               },
               "valor":{
                  "type":"number",
                  "description":"Ammount value",
                  "example":50.9
               }
            }
         },
         "ContaRequest":{
            "required":[
               "idPessoa",
               "limiteSaqueDiario",
               "tipoConta"
            ],
            "type":"object",
            "properties":{
               "idPessoa":{
                  "type":"integer",
                  "description":"Person ID",
                  "format":"int32",
                  "example":3
               },
               "limiteSaqueDiario":{
                  "type":"number",
                  "description":"Withdraw daily limit",
                  "example":1500.99
               },
               "tipoConta":{
                  "type":"string",
                  "description":"Account type. Valid values: PERSONAL or BUSINESS",
                  "example":"PERSONAL",
                  "enum":[
                     "PERSONAL",
                     "BUSINESS",
                     "PERSONAL",
                     "BUSINESS"
                  ]
               }
            }
         },
         "Conta":{
            "type":"object",
            "properties":{
               "idConta":{
                  "type":"integer",
                  "format":"int32"
               },
               "idPessoa":{
                  "type":"integer",
                  "format":"int32"
               },
               "saldo":{
                  "type":"number"
               },
               "limiteSaqueDiario":{
                  "type":"number"
               },
               "flagAtivo":{
                  "type":"boolean"
               },
               "tipoConta":{
                  "type":"integer",
                  "format":"int32"
               },
               "dataCriacao":{
                  "type":"string",
                  "format":"date-time"
               }
            }
         },
         "ResponseConta":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "$ref":"#/components/schemas/Conta"
               }
            }
         },
         "ResponseBigDecimal":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "type":"number"
               }
            }
         },
         "ResponseListTransacao":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/Transacao"
                  }
               }
            }
         }
      }
   }
}